refactor:add Section class for render initialCards
