+++ index.html
@@ -91,6 +91,6 @@
     </figure>
   </li>
 </template>
-<script src="./scripts/index.js" type="module"></script>
+<script src="./pages/index.js" type="module"></script>
 </body>
 </html>
