--- scripts/initialCards.js
@@ -1,26 +1,0 @@
-export const initialCards = [
-  {
-    name: 'Владимир',
-    link: './images/vladimir.jpg'
-  },
-  {
-    name: 'Иваново',
-    link: './images/ivanovo.jpg'
-  },
-  {
-    name: 'Великий Новгород',
-    link: './images/velikiy-novgorod.jpg'
-  },
-  {
-    name: 'Кострома',
-    link: './images/kostroma.jpg'
-  },
-  {
-    name: 'Волгоград',
-    link: './images/volgograd.jpg'
-  },
-  {
-    name: 'Красная Поляна',
-    link: './images/krasnaya-polyana.jpg'
-  }
-];
