+++ utils/constants.js
@@ -1,0 +1,58 @@
+const initialCards = [
+  {
+    name: 'Владимир',
+    link: './images/vladimir.jpg'
+  },
+  {
+    name: 'Иваново',
+    link: './images/ivanovo.jpg'
+  },
+  {
+    name: 'Великий Новгород',
+    link: './images/velikiy-novgorod.jpg'
+  },
+  {
+    name: 'Кострома',
+    link: './images/kostroma.jpg'
+  },
+  {
+    name: 'Волгоград',
+    link: './images/volgograd.jpg'
+  },
+  {
+    name: 'Красная Поляна',
+    link: './images/krasnaya-polyana.jpg'
+  }
+];
+
+const validateSelectors = {
+  formSelector:'.form',
+  inputSelector:'.form__input',
+  buttonSelector: '.form__btn',
+  buttonDisabledClass: 'form__btn_disabled',
+  inputErrorSelector: 'form__input_type_error',
+  spanErrorSelector: 'form__input-error_active'
+}
+
+const popupProfile = document.querySelector('.popup_profile')
+const openPopupProfileBtn = document.querySelector('.profile__btn_edit')
+
+const formElementProfile = popupProfile.querySelector('.form')
+const nameInput = formElementProfile.querySelector('.form__input_name')
+const jobInput = formElementProfile.querySelector('.form__input_job')
+
+const popupPhotoCard = document.querySelector('.popup_photo-card')
+const openPopupPhotoCardBtn = document.querySelector('.profile__btn_add')
+
+const formElementPhotoCard = popupPhotoCard.querySelector('.form')
+const siteInput = formElementPhotoCard.querySelector('.form__input_site')
+const linkInput = formElementPhotoCard.querySelector('.form__input_url')
+
+const profile = document.querySelector('.profile')
+const nameProfile = profile.querySelector('.profile__name')
+const jobProfile = profile.querySelector('.profile__job')
+const photoCardsContainer = document.querySelector('.photo-cards__container')
+
+export {initialCards, validateSelectors,popupProfile, openPopupProfileBtn, formElementProfile,
+nameInput, jobInput,popupPhotoCard, openPopupPhotoCardBtn, formElementPhotoCard,
+siteInput, linkInput,profile,nameProfile, jobProfile, photoCardsContainer}
